fixed issue with refactor
